# docs/openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Inverter Automation API
  version: 1.0.0
  description: >
    Incremental OpenAPI contract baseline used by npm run openapi:check.
    Coverage is still partial; use docs/API_CONTRACT_BASELINE_MAR26.md for the
    current measured backend route inventory until spec parity is complete.
servers:
  - url: /
...
paths:
...
  /api/config:
    get:
      tags: [Config]
      summary: Get current user config
      operationId: getConfig
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Config payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopeConfig"
    post:
      tags: [Config]
      summary: Save current user config
      operationId: saveConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config:
                  type: object
      responses:
        "200":
          description: Save result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopeGeneric"
...
  /api/config/announcement/dismiss:
    post:
      tags: [Config]
      summary: Dismiss a show-once announcement for the current user
      operationId: dismissAnnouncement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: string
                  description: Normalized show-once announcement ID to persist as dismissed
      responses:
        "200":
          description: Dismissal recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelopeAnnouncementDismiss"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelopeBase:
      type: object
      required: [errno, result]
      properties:
        errno:
          type: integer
          description: 0 on success, non-zero for error class.
        result:
          nullable: true
        error:
          type: string
          nullable: true
        meta:
          type: object
          additionalProperties: true
          nullable: true
...
    ApiEnvelopeConfig:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelopeBase"
        - type: object
          properties:
            result:
              type: object
              properties:
                pricingProvider:
                  $ref: "#/components/schemas/PricingProvider"
                amberApiKey:
                  type: string
                  nullable: true
                amberSiteId:
                  type: string
                  nullable: true
                aemoRegion:
                  $ref: "#/components/schemas/AemoRegion"
                siteIdOrRegion:
                  type: string
                  nullable: true
    PricingProvider:
      type: string
      enum: [amber, aemo]
    AemoRegion:
      type: string
      enum: [NSW1, QLD1, VIC1, SA1, TAS1]
      nullable: true
...
    ApiEnvelopeAnnouncementDismiss:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelopeBase"
        - type: object
          properties:
            msg:
              type: string
              nullable: true
            result:
              type: object
              properties:
                id:
                  type: string
                announcementDismissedIds:
                  type: array
                  items:
                    type: string
...
    ApiEnvelopeGeneric:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelopeBase"